sandbox/mimic: Intercept localtime calls
Commit 61be988e5de65aaec4d562db3d9c9c74357f6b01 removed some old,
incomplete code that was intended to close the sandbox service fd.
Although this was technically a fix for the issue that had been raised,
it did not consider the actual purpose of the service, which at the
moment is to handle localtime calls. Now, the mimic zygote will also
intercept these calls, just like upstream.