Remove forcing use of file chooser portal
This reverts commit aeeb6d1508b15716075122e92838a8ccf4787609.

Filters were unfortunately stuck in an unusable state: Chromium and
Electron both use gtk_file_filter_add_custom to add filters, that way
they can perform a case-insensitive comparison. Of course, the callbacks
they pass can't be transferred over the portal, so the filter list ends
up empty.

The only way to work around this would be to extract the std::string
pointer containing the pattern from the callback's user data argument.
However, Chromium uses libc++ and, in particular, its unstable ABI, so
the only way to get the string data from the std::string would be to try
to dynamically figure out what ABI is being used and cast the pointer to
a replica of the proper string. Although this is, in theory, somewhat
doable, it would be incredibly messy, and I'm not confident in its
longetivity or safety.

This *may* return in the future, if I manage to get the above workaround
in place, but for now, the functionality is being removed.