base/evloop: Manually run the sd-event loop
The previous mechanism of using manual polling followed by sd_event_run
would cause issues if sd_event_exit, nothing would be written to the
sd-event fd, resulting in an infinite poll. As a result, this had to
be reworked to manually call sd_event_prepare before polling and
sd_event_wait after polling, that way after Wait() the event loop will
always be in the proper state for a dispatch.

Closes flathub/org.electronjs.Electron2.BaseApp#18.