Refactor code for launches sandboxed processes
This is a massive change, moving the Launcher to base so that the
supervisor can work with the Launcher directly, and thus the
zypak-sandbox processes only forward the argv & fds instead of being
responsible for setting up all the spawn flags and such. This also means
most of the code validating the spawn request can be removed, since the
request is not also setting its own sandbox flags.