Introduce cstring_view as a null-terminated string_view
The code was previously filled with places where we assumed a
string_view was null-terminated, which is non-standard. This was
*workable* because Zypak is small enough that it's easy to guarantee the
source string_view is was null-terminated. However, it's always been
ugly regardless.

This introduces zypak::cstring_view, a string_view-compatible type
that's guaranteed to be null-terminated. In addition, this means that
almost all uses .data() can be removed.